revert routes so dashboard is not home page (yet)
