+++ apps/editor.planx.uk/src/ui/editor/Filter/Filter.tsx
@@ -44,7 +44,7 @@ export const Filters = <T extends object>({
   filterOptions,
 }: FiltersProps<T>) => {
   const navigate = useNavigate();
-  const searchParams = useSearch({ from: "/_authenticated/app/$team/flows" });
+  const searchParams = useSearch({ from: "/_authenticated/app/$team/" });
 
   // Get current filter values from URL
   const getCurrentFilterValues = (): Filters<T> | null => {
